move terrain geometry
